refactor: split tui runtime state
